Move header component from /components/ui to /components
